FK-28-chore 프로젝트 전체 패키징 구조 변경
