--- src/main/java/mono/focusider/application/dto/AuthResponse.java
@@ -1,19 +1,0 @@
-package mono.focusider.application.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AuthResponse {
-    private String accessToken;
-    private String refreshToken;
-
-    public AuthResponse() {
-    }
-
-    public AuthResponse(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
