--- src/main/java/mono/focusider/application/dto/LoginRequest.java
@@ -1,20 +1,0 @@
-package mono.focusider.application.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class LoginRequest {
-    @NotBlank(message = "Username is required")
-    @JsonProperty("username")
-    private String username;
-
-    @NotBlank(message = "Password is required")
-    @JsonProperty("password")
-    private String password;
-}
