--- src/main/java/mono/focusider/application/dto/SignupRequest.java
@@ -1,28 +1,0 @@
-package mono.focusider.application.dto;
-
-import java.time.LocalDate;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Past;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class SignupRequest {
-    @NotBlank(message = "Username is required")
-    private String username;
-
-    @NotBlank(message = "Password is required")
-    private String password;
-
-    @NotBlank(message = "Name is required")
-    private String name;
-
-    @NotNull(message = "Birth date is required")
-    @Past(message = "Birth date must be in the past")
-    private LocalDate birthDate;
-}
