--- src/main/java/mono/focusider/application/dto/UsernameAvailabilityResponse.java
@@ -1,17 +1,0 @@
-package mono.focusider.application.dto;
-
-public class UsernameAvailabilityResponse {
-    private boolean available;
-
-    public UsernameAvailabilityResponse(boolean available) {
-        this.available = available;
-    }
-
-    public boolean isAvailable() {
-        return available;
-    }
-
-    public void setAvailable(boolean available) {
-        this.available = available;
-    }
-}
