+++ src/main/java/mono/focusider/domain/auth/dto/AuthResponseDto.java
@@ -1,0 +1,19 @@
+package mono.focusider.domain.auth.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AuthResponseDto {
+    private String accessToken;
+    private String refreshToken;
+
+    public AuthResponseDto() {
+    }
+
+    public AuthResponseDto(String accessToken, String refreshToken) {
+        this.accessToken = accessToken;
+        this.refreshToken = refreshToken;
+    }
+}
