+++ src/main/java/mono/focusider/domain/auth/dto/LoginRequestDto.java
@@ -1,0 +1,20 @@
+package mono.focusider.domain.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class LoginRequestDto {
+    @NotBlank(message = "Username is required")
+    @JsonProperty("username")
+    private String username;
+
+    @NotBlank(message = "Password is required")
+    @JsonProperty("password")
+    private String password;
+}
