+++ src/main/java/mono/focusider/domain/auth/dto/UsernameAvailabilityResponseDto.java
@@ -1,0 +1,17 @@
+package mono.focusider.domain.auth.dto;
+
+public class UsernameAvailabilityResponseDto {
+    private boolean available;
+
+    public UsernameAvailabilityResponseDto(boolean available) {
+        this.available = available;
+    }
+
+    public boolean isAvailable() {
+        return available;
+    }
+
+    public void setAvailable(boolean available) {
+        this.available = available;
+    }
+}
