+++ src/main/java/mono/focusider/domain/member/domain/User.java
@@ -1,0 +1,20 @@
+package mono.focusider.domain.member.domain;
+
+import java.time.LocalDate;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private LocalDate birthDate;
+    private String username;
+    private String password;
+}
