--- src/main/java/mono/focusider/domain/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package mono.focusider.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import mono.focusider.domain.model.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-
-    boolean existsByUsername(String username);
-}
