+++ src/main/java/mono/focusider/global/Security/JwtAuthenticationProvider.java
@@ -1,0 +1,38 @@
+package mono.focusider.global.Security;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
+public class JwtAuthenticationProvider implements AuthenticationProvider {
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
+    @Override
+    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+        String token = (String) authentication.getCredentials();
+        String username = jwtTokenUtil.getUsernameFromToken(token);
+
+        if (username != null && jwtTokenUtil.validateToken(token)) {
+            List<GrantedAuthority> authorities = new ArrayList<>();
+            return new UsernamePasswordAuthenticationToken(username, null, authorities);
+        }
+
+        throw new BadCredentialsException("Invalid token");
+    }
+
+    @Override
+    public boolean supports(Class<?> authentication) {
+        return (UsernamePasswordAuthenticationToken.class.isAssignableFrom(authentication));
+    }
+}
