+++ src/main/java/mono/focusider/global/config/OpenApiConfig.java
@@ -1,0 +1,18 @@
+package mono.focusider.global.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfig {
+
+    @Bean
+    public OpenAPI usersMicroserviceOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("Auth Service API")
+                        .description("Your API Description")
+                        .version("1.0"));
+    }
+}
