+++ src/main/java/mono/focusider/global/exception/InvalidRefreshTokenException.java
@@ -1,0 +1,16 @@
+package mono.focusider.global.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class InvalidRefreshTokenException extends RuntimeException {
+
+    public InvalidRefreshTokenException(String message) {
+        super(message);
+    }
+
+    public InvalidRefreshTokenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
