--- src/main/java/mono/focusider/infrastructure/Security/JwtAuthenticationFilter.java
@@ -1,99 +1,0 @@
-package mono.focusider.infrastructure.Security;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-import java.security.Key;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtSecretKey jwtSecretKey;
-
-    public JwtAuthenticationFilter(JwtSecretKey jwtSecretKey) {
-        this.jwtSecretKey = jwtSecretKey;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
-        if (isPublicEndpoint(request)) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        Cookie[] cookies = request.getCookies();
-        if (cookies != null) {
-            Optional<Cookie> jwtCookie = Arrays.stream(cookies)
-                    .filter(cookie -> "access_token".equals(cookie.getName()))
-                    .findFirst();
-
-            if (jwtCookie.isPresent()) {
-                String token = jwtCookie.get().getValue();
-                try {
-                    UsernamePasswordAuthenticationToken authToken = validateTokenAndGetAuthentication(token);
-                    SecurityContextHolder.getContext().setAuthentication(authToken);
-                } catch (Exception e) {
-                    response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-                    return;
-                }
-            } else {
-                response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-                return;
-            }
-        } else {
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private boolean isPublicEndpoint(HttpServletRequest request) {
-        String path = request.getRequestURI();
-        return path.startsWith("/api/auth/") || path.startsWith("/swagger-ui") || path.startsWith("/api-docs")
-                || path.startsWith("/v3/api-docs");
-    }
-
-    private UsernamePasswordAuthenticationToken validateTokenAndGetAuthentication(String token) {
-        try {
-            Key signingKey = Keys.hmacShaKeyFor(jwtSecretKey.getSecretKey().getBytes());
-
-            Claims claims = Jwts.parserBuilder()
-                    .setSigningKey(signingKey)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-
-            String username = claims.getSubject();
-            @SuppressWarnings("unchecked")
-            List<String> authorities = claims.get("authorities", List.class);
-
-            UserDetails userDetails = new User(username, "",
-                    authorities.stream()
-                            .map(SimpleGrantedAuthority::new)
-                            .collect(Collectors.toList()));
-
-            return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-        } catch (Exception e) {
-            throw new RuntimeException("Invalid token", e);
-        }
-    }
-}
