--- src/main/java/mono/focusider/infrastructure/config/OpenApiConfig.java
@@ -1,18 +1,0 @@
-package mono.focusider.infrastructure.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiConfig {
-
-    @Bean
-    public OpenAPI usersMicroserviceOpenAPI() {
-        return new OpenAPI()
-                .info(new Info().title("Auth Service API")
-                        .description("Your API Description")
-                        .version("1.0"));
-    }
-}
