--- src/main/java/mono/focusider/infrastructure/exception/GlobalExceptionHandler.java
@@ -1,21 +1,0 @@
-package mono.focusider.infrastructure.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<String> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        log.info("Validation error: " + ex.getMessage());
-        return new ResponseEntity<>("Validation error: " + ex.getBindingResult().getAllErrors(),
-                HttpStatus.BAD_REQUEST);
-    }
-}
